LayoutTests/imported/w3c/web-platform-tests/webcodecs/temporal-svc-encoding.https.any_h264-expected.txt

CONSOLE MESSAGE: Error: assert_own_property: expected property "svc" missing
CONSOLE MESSAGE: Error: assert_own_property: expected property "svc" missing

Harness Error (FAIL), message = Error: assert_own_property: expected property "svc" missing

FAIL SVC L1T2 assert_equals: expected 24 but got 1
FAIL SVC L1T3 assert_equals: expected 24 but got 1
